Move timeout/window_size to model
