+++ src/channel/channel.rs
@@ -6,7 +6,7 @@ use crate::constant::ssh_msg_code;
 use crate::error::{SshError, SshResult};
 use crate::model::Data;
 use crate::slog::log;
-use crate::window_size::WindowSize;
+use crate::model::WindowSize;
 use std::cell::RefCell;
 use std::rc::Rc;
 use std::{
