+++ src/model/mod.rs
@@ -1,5 +1,9 @@
 pub(crate) mod data;
 pub(crate) mod packet;
+pub(crate) mod timeout;
+pub(crate) mod window_size;
 
 pub(crate) use data::Data;
 pub(crate) use packet::Packet;
+pub(crate) use timeout::Timeout;
+pub(crate) use window_size::WindowSize;
